TB-16

+++ src/services/home.service.js
@@ -4,6 +4,7 @@ import AppError from "../errors/AppError.js";
 import Home from "../models/home.model.js";
 import mongoose from "mongoose";
 import Room from "../models/room.model.js";
+import HomeOwner from "../models/homeOwner.model.js";
 const inserHomeIntoDB = async (payload) => {
   const result = await Home.create(payload);
   return result;
@@ -24,16 +25,16 @@ const getAllHomes = async (query) => {
 };
 
 const getSingleHome = async (id) => {
-  const result = await Home.findOne({ _id: id }).populate("user");
+  const result = await Home.findOne({ _id: id });
   return result;
 };
 const updateHome = async (id, payload) => {
   const result = await Home.findByIdAndUpdate(id, payload, {
     new: true,
-  }).populate("user");
+  });
   return result;
 };
-const deleteHome = async (id) => {
+const deleteHome = async (homeOwnerId, id) => {
   const session = await mongoose.startSession();
   try {
     session.startTransaction();
@@ -44,8 +45,30 @@ const deleteHome = async (id) => {
         { session })
       );
     }
-    await Room.deleteMany({ home: id }, { session });
-
+    const updateHomeOwner = await HomeOwner.findOneAndUpdate(
+      {
+        id: homeOwnerId,
+      },
+      {
+        $pull: {
+          homes: deleteHome[0]?._id,
+        },
+      },
+      { new: true, session }
+    );
+    if (!updateHomeOwner) {
+      throw new AppError(
+        httpStatus.BAD_REQUEST,
+        "Failed To Delete Home.Please try again"
+      );
+    }
+    const deleteRoom = await Room.deleteMany({ home: id }, { session });
+    if (!deleteRoom.acknowledged) {
+      throw new AppError(
+        httpStatus.BAD_REQUEST,
+        "Failed To Delete Home.Please try again"
+      );
+    }
     await session.commitTransaction();
     await session.endSession();
     return deleteHome;
